h2overlord-python/pyproject.toml
[project]
name = "h2overlord-python"
version = "0.1.0"
description = ""
authors = [
    {name = "erdeanmich",email = "erdeanlinda@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "gpiozero (>=2.0.1,<3.0.0)",
    "rpi-gpio (>=0.7.1,<0.8.0) ; sys_platform == \"linux\"",
    "rpi-bme280 (>=0.2.4,<0.3.0)",
    "schedule (>=1.2.2,<2.0.0)",
    "pendulum (>=3.1.0,<4.0.0)",
    "bottle (>=0.13.3,<0.14.0)",
]

[tool.poetry]
packages = [{include = "h2overlord_python", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
